Broker DB: quit if conn fails, flyway migration; log4j configs
Before this commit, the broker would simply switch to an in-memory DB
if the configured database was not reachable.  Now we fail instead to
not lose precious data.  Refactored DB code while I was on it.

Now uses flyway migration, was way easier than I thought.

broker and architect have log4j configs which log on the INFO level.